requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Django==4.2.4
django-admin-rangefilter==0.8.5
django-baton==2.8.0
django-bulk-update==2.2.0
django-celery-beat==2.1.0
django-ckeditor==6.3.2
django-cors-headers==3.10.1
django-cors-middleware==1.5.0
django-filter==23.1
django-js-asset==2.0.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.2
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.25.2
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==2.1.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
